reflection fix

+++ includes/structures.h
@@ -6,7 +6,7 @@
 /*   By: pgritsen <pgritsen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/24 13:21:23 by pgritsen          #+#    #+#             */
-/*   Updated: 2018/02/27 15:39:43 by pgritsen         ###   ########.fr       */
+/*   Updated: 2018/02/28 18:40:01 by pgritsen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 typedef struct	s_light
 {
 	char	type;
-	double	intens;
+	float	intens;
 	t_point	pos;
 	t_point	dir;
 }				t_light;
@@ -27,16 +27,16 @@ typedef struct	s_obj
 	short int	type;
 	t_point		pos;
 	t_point		dir;
-	double		rad;
+	float		rad;
 	int			spec;
-	double		refl;
+	float		refl;
 }				t_obj;
 
 typedef struct	s_viewport
 {
-	double	w;
-	double	h;
-	double	dist;
+	float	w;
+	float	h;
+	float	dist;
 }				t_viewport;
 
 #endif
